requirements.txt
Pillow
numpy
PySide6